requirements.txt
discord.py==2.5.2
nest_asyncio
pytz
apscheduler
Pillow